Address warnings

+++ app/src/main/java/com/island/androidsftpdocumentsprovider/provider/CleanOldCache.kt
@@ -12,17 +12,33 @@ class CleanOldCache : BroadcastReceiver() {
     val TAG = "CleanOldCache"
 
     override fun onReceive(context: Context, intent: Intent?) {
+        if (intent?.action != Intent.ACTION_MY_PACKAGE_REPLACED) {
+            Log.e(TAG, "Ignoring action ${intent?.action}")
+            return;
+        }
+
         Log.d(TAG, "Cleaning old cache")
         val cacheDir = context.getCacheDir();
-        for(subdir in cacheDir.listFiles()) {
+        if(cacheDir == null) {
+            Log.e(TAG, "Cachedir is null")
+            return
+        }
+        val subdirs = cacheDir.listFiles()
+        if(subdirs == null) {
+            Log.e(TAG, "No cache subdirs")
+            return
+        }
+        for(subdir in subdirs) {
             if(subdir.name.length < 44) {
                 Log.d(TAG, "Skipping "+subdir.name)
                 continue;
             }
-            for(file in subdir.listFiles()) {
-                Log.i(TAG, "Removing file "+file)
-                file.delete()
-            }
+            val files = subdir.listFiles()
+            if(files != null)
+                for(file in files) {
+                    Log.i(TAG, "Removing file "+file)
+                    file.delete()
+                }
             subdir.delete()
             Log.i(TAG, "Removing subdirectory "+subdir)
         }
